Increase resolution of RDF.
We were using a very low resolution in all the radial distribution functions.
We used to do this because the old MDFT was quite noisy and unstable.
Thus, making low resolution histograms was a way of smoothing the data.
Nowadays, since MDFT is much more robust and stable, such smoothing
is a loss of information.

With this commit, the width of a histogram is 0.01 Angstrom.
It was something like 0.1 or even lower before, and was decided upon Sturges rules.
This is not the case anymore: 0.1 Ang is a good compromise, from my testings.