Dont print (non-usefull) zeros in the RDF.
dontPrintZeros.
We have empty bins in the RDF:
 - the ones in the core of the solute, for which we want to print rdf(x in core)=0.
 - the ones that are empty because of lack of information, for x > beginning of the first peak.
These last ones we dont want to print.
We thus first print the zero, but as soon as we detect the rdf starts to be nonzero,
we dont print the zeros anymore.

I see one drawback:
The number of plotted values can not be predicted and changes with the system one studies.
This may lead to some complications in specific postprocessing outside MDFT.
For instance, it is easy to substract two sets of values in xmgrace if they have same size and abscissa.
If they don't, you'll have to find a way, for instance using interpolation and picking on each of the set
to build two new sets with the same sizes. That's a really specific drawback and don't lower the interest
in this feature, to my opinion.