File generated by output_rdf now in xmgrace format.
Output_rdf generates a two-column file containing the site site
radial distribution functions.

With this patch, we produce these files with few lines of header
starting with symbol @ that xmgrace recognizes.
For instance, we now have legends, xlabel and ylabel to plotted when
xmgrace output/rdf.xvg

output/rdf.xvg replaces output/rdf.out

xvg files are ascii files with few lines of header on top of the data.
Nothing to be scared about.

This is also the format that gromacs uses (gmx energy to be precise).