Move back BPP files to f90
That's the first step in the move toward a BPP free MDFT ...
Right now this only rename the files.
+
Inquire if file exist before counting lines

When calling n_linesinfile, we expected the user to provide filenames that exist.
We now check the existence of the file before trying to open it