Use a temporary array for c with interpolation
We were using c_(imnmunukhi,iq) with iq taken from the
list of points of Luc.
No we use a single temp array ceff, where eff stands for
effective.
This ceff(1:na) is a linear interpolation of c(:,iq) and c(:,iq+1).
It is a bit of supplementary work inside the xyz loop, BUT
at the end it is more stable and I end up with fewer iterations.
For instance for CH4+1, I have 24 iterations instead of 25 to converge to 1E-5.
Thus, it is faster !