Build FFTW-3.3.4 if needed automatically, by default
You may disable this behaviour by  -DBUILD_FFTW=OFF   since the new default is -DBUILD_FFTW=ON.
If FFTW is found on the computer by FindPackage, then it is used.
If it is not found, the default is to build FFTW-3.3.4.
If it is not found and -DBUILD_FFTW=OFF, then a fatal error happens with an explicit message.