Do again the pressure correction (PC/PC+)
Was commented for a long time due to changes in the data structures.
Rewritten and tested on CH4.
Corrections are typically of the order of 30 kJ/mol for unified methane.

ISc (also called PC) correction is completely justified.
ISc* (also called PC+) correction has no justification but that it works well.

See papers by Volodymyr, Levesque and Borgis around 2014.