Add benchmarks small/mid/big input files
In issue #95, @youldrouis asks for representative benchmarks.
He needs input files that are representative of real applications
If applied, this commit will add such examples input files.

For the small case, I was able to do it in a rush on my laptop in few minutes.
The results are found in result.stdout.