Fix Test 4 of histograms
Test 4 was a crazy one. There was no reason it would work. The point is the bug would popup for very large number of grid nodes. With this patch, we remove all unit testing. We know this routines work. No reason to do such  extensive unit testing at each run.