Add optional acceptable values to getinput%char
Check that the string entered by the user is one of the valid values.
The string entered by the user is in "input_char".
The input tag string looked for by the program is "tag".
Acceptable, valid, values are optional. If present, they are in the array of strings "validValues".

getinput%char now has a new optional argument: validValues.
validValues is an array of strings of implicit length. It should be used as follow :
```fortran
getinput%char( "tag" , validValues = [ "toto", "tata", "titi"]
```

Because it is Fortran and because I'm lazy, all strings of the array validValues must have the same length. Thus, you will be required to add empty spaces at the end of all but the lengthiest string. For instance :
```fortran
getinput%char( "tag" , validValues = [ "tot ", "ta  ", "titi"]
```