Add the benchmark minimizer mode
With this commit, a new tag appears in dft.in : `minimizer`.
`minimizer` can be one of `sd`, `lbfgs` or `benchmark`.
By default, if the tag `minimizer` is absent, the lbfgs minimizer is used.
`minimizer = sd` should be prefered over the old `minimize_with_steepest_descent = T`.
In order not to break all the scripts, especially @cgageat's , I  kept the old tag `minimize_with_steepest_descent`.  I'll open an issue with the removing of this deprecated tag.

If `minimizer = benchmark` is used, then MDFT works as usual for everything, except that no minimization is done.
In this mode, MDFT does `maximum_iteration_nbr` (default = 30) evaluations of the functional and its gradient without any modification of the density field. All the evaluations are exactly the same. This mode has several advantages :
- all benchmarks will end up with the same number of evaluations of the functional. We won't have this source of differences where on one computer you could have 30 evaluations but 31 on another computer.
- all outputs should always be the same.
- the spatial grid resolution can be as low as one wants, for instance to study the influence of mmax vs nx*ny*nz.

This commit consequently also updates the example input files.