Change convergence criteria to 0.01 kJ/mol
0.01 kJ/mol is more precise than what we can expect from our theory, largely.
Thus, there is no need to converge to 0.00001 kJ/mol.
Furthermore, we were converging to a fraction (very low value) of the evaluation of the functional.
This was non-sense since we are studying objects with solvation free energies that differ by orders
of magnitude, like large charged proteins vs methane.