Dont print radial polarization for complex solutes
since it is not readable to have radial distribution functions of non spherical symetries.
Moreover, Yacine pointed out that it takes time for complex solutes, for instance several
tens of seconds in the case of proteins with 4000 sites.

The choice of plotting them only for solutes of 10 sites is arbitrary and comes from
my experience (Max).