Temporary fix for stopping after N iterations
Stop the execution if the program reaches a number of iteration that has been
tagged by the user as a maximum before stopping.

We have a tag, maximum_iteration_nb,  to stop the convergence. After this nb
of iteration, the code goes to the postprocessing etc.

We the new tag, maximum_iteration_before_error, the code stops the whole
execution at this number of iteration.

This is a temporary fix for issue #143
This is temporary since no other minimizer than LBFGSB (our default)
implement this fix.

I (Max) did this for Sohvi to be able to go further in her last week
before the holidays.