Input file name defined by user only once, not twice anymore
In order to change the input file name, dft.in in MDFT, to anything like hugo.in, the user had to search for dft.in
in the whole module, and to change it at two locations. The second location is now replaced by a smart :P use of
the first declaration.